requirements.txt
langchain_google_genai
langchain
dotenv
pillow
transformers
torch
diffusers 
accelerate
huggingface_hub
langchain-tavily
BeautifulSoup
selenium
instagrapi